# net/netip: optimize parseIPv4 and refactor IPv6 embedded IPv4 parsing